pyproject.toml
[project]
name = "medical-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.61",
    "langchain-experimental>=0.3.4",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=2.1.4",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
